moved catalog test into another test case
